스프링 DB 2편 - 데이터 접근 활용 기술/springtx/src/test/java/hello/springtx/propagation/MemberServiceTest.java
package hello.springtx.propagation;

import lombok.extern.slf4j.Slf4j;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static org.assertj.core.api.Assertions.assertThat;

@Slf4j
@SpringBootTest
class MemberServiceTest {
    @Autowired
    MemberService memberService;
    @Autowired
    MemberRepository memberRepository;
    @Autowired
    LogRepository logRepository;

    /**
     * MemberService    @Transactional:OFF
     * MemberRepository @Transactional:ON
     * LogRepository    @Transactional:ON
     */
    @Test
    void outerTxOff_success() {
        // given
        var username = "outerTxOff_success";

        // when
        this.memberService.joinV1(username);

        // then : 모든 데이터가 정상 저장된다.
        assertThat(this.memberRepository.find(username).isPresent()).isTrue();
        assertThat(this.logRepository.find(username).isPresent()).isTrue();
    }
}